pyproject.toml
[tool.poetry]
name = "api-posting-system"
version = "0.1.0"
description = ""
authors = ["lazzarusxd <joaolazarofera@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
cryptography = "^44.0.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.10.4"}
psycopg2-binary = "^2.9.10"
python-jose = "^3.3.0"
pytz = "^2024.2"
httpx = "^0.28.1"
python-multipart = "^0.0.20"
python-dotenv = "^1.0.1"
aio-pika = "^9.5.4"
aiosmtplib = "^3.0.2"
redis = "^5.2.1"
asyncpg = "^0.30.0"
brazilcep = "^6.7.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-mock = "^3.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
